+++ packages/pwa-kit-extension-sdk/src/configs/webpack/overrides-resolver-loader.ts
@@ -5,20 +5,35 @@
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 import {LoaderContext} from 'webpack'
+import fse from 'fs-extra'
 import os from 'os'
 import path from 'path'
 import resolve from 'resolve'
 
 // Local Imports
-import {buildCandidatePaths, getPackageName} from '../../shared/utils'
+import {buildCandidatePaths, getPackageName, SETUP_FILE_REGEX} from '../../shared/utils'
 
 // Types
 import type {ExtendedCompiler} from './types'
 
 // Constants
+const EXTENSION_PACKAGE_PREFIX = 'extension-'
+const EXTENSION_PACKAGE_NAMESPACE = '@salesforce'
 const IMPORT_REGEX = /import\s+(?:(?:[\w*\s{},]*)\s+from\s+)?['"](\..*?)['"]/g
+const OVERRIDABLE_FILE_NAME = '.force_overrides'
+const MONO_REPO_WORKSPACE_FOLDER = 'packages'
+const NODE_MODULES_FOLDER = 'node_modules'
 const REQUIRES_REGEX = /require\(['"](\..*?)['"]\)/g
 const SRC = 'src'
+
+// Cache for processed overrides
+const OVERRIDABLE_CACHE = {
+    node: [] as string[],
+    web: [] as string[]
+}
+
+// Export the cache for testing purposes
+export const __OVERRIDABLE_CACHE__ = OVERRIDABLE_CACHE
 
 /**
  * Webpack loader to override the resolution of a module based on the PWA-Kit applications
@@ -54,8 +69,8 @@ const OverrideResolverLoader = function (this: LoaderContext<any>) {
     }
 
     // Lets use the compiler configuration to ensure we are resolving the correct file extensions.
-    const compilerOptions = this._compiler!.options
-    const extensions = compilerOptions.resolve?.extensions || []
+    const compilerOptions = this._compiler?.options
+    const extensions = compilerOptions?.resolve?.extensions || []
     const basedir = options?.baseDir || process.cwd()
     const applicationExtensions = compiler?.custom?.extensions || []
 
@@ -112,6 +127,110 @@ const OverrideResolverLoader = function (this: LoaderContext<any>) {
         // future dev's this might be a place that needs to be adjusted.
         callback(null, adjustedSource)
     })
+}
+
+/**
+ * Return a boolean indicating if the source file should be processed by the override loader based on
+ * various conditions including the cache state, the source file type, and the presence of an override file
+ * in the provided overridables list.
+ *
+ * @param source
+ * @param options
+ * @param {string} [options.projectDir] - Loader-specific options.
+ * @param {string} [options.target=DEFAULT_TARGET] - The target environment, either 'node' or 'web'.
+ * @param {boolean} [options.isMonoRepo] - Changes the source path normalization based on if the project is run in a mono-repo.
+ * @returns {boolean} - A boolean indicating if the source file should be processed by the override loader.
+ */
+export const validateOverrideSource = (source: string, options: any = {}) => {
+    const {isMonoRepo = false, target = 'node', overridables = []} = options
+    const isExtensionFile = source.includes(`${path.sep}${EXTENSION_PACKAGE_PREFIX}`)
+    const isSetupFile = SETUP_FILE_REGEX.test(source)
+    const targetCache = OVERRIDABLE_CACHE[target as keyof typeof OVERRIDABLE_CACHE]
+
+    // Exit early if we have:
+    // 1. Processed this file already.
+    // 2. The file is not an extension file.
+    // 3. The file is an extension setup file.
+    if (targetCache.includes(source) || !isExtensionFile || isSetupFile) {
+        return false
+    }
+
+    // Because our webpack configuration is setup to resolve symlinks, we need to normalize the source path because
+    // the source path passed to the loaded is not representative of what you would see in a generated project (e.g.
+    // it doesn't resolve to being in the node_modules folder).
+    let normalizedSource = ''
+
+    // We are only concerned with the source path relative to the extension package namespace.
+    normalizedSource = `${
+        source
+            .split(
+                `${path.sep}${isMonoRepo ? MONO_REPO_WORKSPACE_FOLDER : NODE_MODULES_FOLDER}${
+                    path.sep
+                }`
+            )
+            .pop() ?? ''
+    }`
+
+    // At this point the path is either POSIX or windows, we need to normalize it to POSIX.
+    normalizedSource = normalizedSource.replace(/\\/g, '/')
+
+    // NOTE:
+    // For now we are going to make the assumption that all the extension projects in our mono repo
+    // are part of the `@salesforce` namespace, this is pretty safe. So we are going to add the namespace.
+    normalizedSource = `./${NODE_MODULES_FOLDER}/${
+        isMonoRepo ? EXTENSION_PACKAGE_NAMESPACE + path.posix.sep : ''
+    }${normalizedSource}`
+
+    // Check if the normalized source is in the list of overridables.
+    const hasOverride = overridables.includes(normalizedSource)
+
+    // If we have an override, add it to the cache so we don't process it again.
+    if (hasOverride) {
+        targetCache.push(source)
+    }
+
+    return hasOverride
+}
+
+/**
+ * Generates a Webpack rule for application extensibility, configuring the loader for
+ * handling application extensions based on the target (e.g., 'node' for server-side,
+ * 'web' for client-side). Specifically this enables the ability to override the resolution
+ * via the `.force_overrides` file.
+ *
+ * @param {Object} [options={}] - Options to customize the Webpack rule.
+ * @param {string} [options.projectDir] - Loader-specific options.
+ * @param {string} [options.target=DEFAULT_TARGET] - The target environment, either 'node' or 'web'.
+ * @param {boolean} [options.isMonoRepo] - Changes the source path normalization based on if the project is run in a mono-repo.
+ *
+ * @returns {Object} A Webpack rule configuration object for handling application extensions.
+ */
+export const ruleForOverrideResolver = (options: any = {}) => {
+    const {projectDir, target, isMonoRepo} = options
+    let overridables: string[] = []
+
+    try {
+        overridables = fse
+            .readFileSync(path.join(projectDir, OVERRIDABLE_FILE_NAME), 'utf8')
+            .split(/\r?\n/)
+            .filter((line) => !line.startsWith('//'))
+    } catch (e) {
+        // If where is no .force_overrides file, we can safely return null.
+        return null
+    }
+
+    return {
+        test: (source: string) => {
+            return validateOverrideSource(source, {
+                isMonoRepo,
+                target,
+                overridables
+            })
+        },
+        use: {
+            loader: '@salesforce/pwa-kit-extension-sdk/configs/webpack/overrides-resolver-loader'
+        }
+    }
 }
 
 // Export the loader as the default export with proper typing