+++ packages/pwa-kit-dev/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## v4.0.0-extensibility-preview.3 (Jan 14, 2025)
 - Ensure build command works on windows when path includes spaces. [#2204](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2204)
+- Add loader rule for `override-resolver-loader`. [#2207](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2207)
 ## v4.0.0-extensibility-preview.2 (Dec 09, 2024)
 ## v4.0.0-extensibility-preview.1 (Dec 09, 2024)
 ## v4.0.0-extensibility-preview.0 (Nov 28, 2024)